+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "t2-streamlit"
-version = "0.6.2"
+version = "0.7.0"
 description = "Streamlit app for AI control system"
 authors = [
     {name = "j25ng", email = "jsan2861@gmail.com"},
